Replace AMSR2 logic
Completely refactor the logic to process AMSR 2 Lance SWE data.

SWE file will be warped into the MODIS projection and clipped to the that
extent. Afterwards the overlap of MODIS tiles with a SWE tile will be
determined and the SWE value distributed based on percentage of MODIS SCA
percentages.